When handle SINGLE message, never send them back
When a node appears and immediately reappears there is a small period of
time that the node list can be out of sync. This can cause loops in the
mesh (A thinks we have to go through B to get to C, while B thinks we
have to go back through A). This change make sure B never sends back to
A and therefore prevents loops from occuring during this period.